+++ changelog.md
@@ -2,6 +2,11 @@ This file contains a summary of changes to Haskell.nix and `nix-tools`
 that will impact users.
 
 ## July 21, 2020
+* Removed `components.all`, use `symlinkJoin` on components.exes or
+ `shellFor` if you need a shell.
+* Added `components` arguemnt to `shellFor`.
+
+## July 21, 2020
 * Added GHC 8.8.4 and replaced 8.8.3 in tests and as the ghc
   used to build nix-tools for stack projects.